Revert defaultPasswordValidator change
