--- packages/ui/src/helpers/shared/index.ts
@@ -1,1 +1,0 @@
-export * from './validator';
